Moved OperationTracker functionality to the the internal page since this isn't public API
