+++ hazelcast/src/main/java/com/hazelcast/spi/impl/operationservice/LiveOperations.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.spi;
+package com.hazelcast.spi.impl.operationservice;
 
 import com.hazelcast.nio.Address;
